LayoutTests/fast/events/autoscroll-with-software-keyboard-expected.txt

This test presses return until the cursor would be behind the keyboard, forcing the content to be scrolled to stay in view with a software keyboard on screen.
PASS: page has scrolled when caret is behind software keyboard.
PASS: page has continuted to scroll when caret is beyond the edge of the screen.
